+++ CosmosDBShell/Azure.Data.Cosmos.Shell.Util/ShellLiteral.cs
@@ -1,0 +1,59 @@
+// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.  All rights reserved.
+// ------------------------------------------------------------
+
+namespace Azure.Data.Cosmos.Shell.Util;
+
+using System.Globalization;
+using System.Text;
+
+internal static class ShellLiteral
+{
+    public static string Quote(string? value)
+    {
+        value ??= string.Empty;
+
+        var builder = new StringBuilder(value.Length + 2);
+        builder.Append('"');
+
+        foreach (char character in value)
+        {
+            switch (character)
+            {
+                case '"':
+                    builder.Append("\\\"");
+                    break;
+                case '\\':
+                    builder.Append("\\\\");
+                    break;
+                case '$':
+                    builder.Append("\\$");
+                    break;
+                case '\n':
+                    builder.Append("\\n");
+                    break;
+                case '\r':
+                    builder.Append("\\r");
+                    break;
+                case '\t':
+                    builder.Append("\\t");
+                    break;
+                default:
+                    if (char.IsControl(character))
+                    {
+                        builder.Append("\\u");
+                        builder.Append(((int)character).ToString("X4", CultureInfo.InvariantCulture));
+                    }
+                    else
+                    {
+                        builder.Append(character);
+                    }
+
+                    break;
+            }
+        }
+
+        builder.Append('"');
+        return builder.ToString();
+    }
+}